+++ lib/src/utils/loader_button_enum_states.dart
@@ -1,1 +1,0 @@
-enum LoaderButtonEnumStates { initial, loading, reset }